Solr 9.4 → 9.5 for integration tests
