+++ tests/Integration/Proxy/CustomizedCurlTest.php
@@ -11,6 +11,7 @@ use Solarium\Tests\Integration\TestClientFactory;
  * Test connecting through a proxy with a customized Curl adapter that sets the proxy options differently.
  *
  * @group integration
+ * @group proxy
  */
 class CustomizedCurlTest extends CurlTest
 {
